requirements.txt
aiohttp==3.7.3
alembic==1.5.4
appnope==0.1.2
argon2-cffi==20.1.0
arrow==0.17.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
aws-lambda-builders==0.6.0
aws-sam-cli==0.33.1
aws-sam-translator==1.15.1
backcall==0.2.0
BentoML==0.11.0
binaryornot==0.4.4
bleach==3.3.0
boto3==1.17.7
botocore==1.20.7
Cerberus==1.3.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
chevron==0.14.0
click==7.1.2
configparser==5.0.1
cookiecutter==1.6.0
cycler==0.10.0
dateparser==0.7.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
docker==4.2.2
entrypoints==0.3
Flask==1.0.4
future==0.18.2
grpcio==1.35.0
gunicorn==20.0.4
humanfriendly==9.1
idna==2.8
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.9.5
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
ply==3.11
poyo==0.5.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
py-zipkin==0.20.1
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.0
python-editor==1.0.4
python-json-logger==2.0.1
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2020.11.13
requests>=2.22.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
scikit-learn==0.24.1
scipy==1.6.0
Send2Trash==1.5.0
serverlessrepo==0.1.9
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.7
tabulate==0.8.7
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
thriftpy2==0.4.14
tomlkit==0.5.8
torch==1.6.0
torchvision==0.7.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
yarl==1.6.3
zipp==3.4.0
